requirements.txt
langchain-community
langchain-huggingface
transformers
reportlab
ddgs
torch
langchain-ollama
langchain-tavily
langchain-openai
arxiv==1.4.8
sentence-transformers
faiss-cpu
unstructured
pdfminer
pypdfium2
pdfminer.six==20220524
trafilatura
langchain-google-genai
pymupdf